Move upgrade button to async-loaded component
